fix: move new model file functionality into editor actions
